[patches] fix freifunk-policyrouting
At the moment netifd can not create ip rules for wirless interfaces or only some
rules. This is due to the async nature of wireless interfaces and how netifd
initializes these interfaces.

This change makes sure that the unreachable rule is created even if there is
already a rule for the olsr-default table lookup. The unreachable rule does need
a condition on its own and should not be coverd by the olsr-default table lookup
condition in the if statement.