[patches] pr - add rule for tunl0 to olsr-tunnel
The default route of VPN03 will be added to olsr-tunnel table. tunl0 is the
interface for incoming SmartGateway tunnels. As there was no rule for tunl0 to
look into olsr-tunnel routing table all SmartGateway clients had no uplink
access. This changeset adds an additional ip rule for tunl0 for the olsr-tunnel
table like "19999:	from all iif tunl0 lookup olsr-tunnel".

Normally this should not be necessary because the hotplug.d policyrouting script
should take care of this but netifd does not trigger an event for tunl0 (tunnel
interfaces) and therefor this script will never be called.