[configs] Remove ssp support and strip kernel
Stack smashing protection got merged with 6318d5a. As there is no package which
supports SSP at the moment in the Freifunk Firmware Berlin this functionality
is obsolete and can be deleted for size and complexity reasons (see
https://github.com/freifunk-berlin/firmware/pull/107 and
https://github.com/freifunk-berlin/firmware/issues/191#issuecomment-68630029)

Activate CONFIG_STRIP_KERNEL_EXPORTS for size reasons. OpenWRT supports
stripping unnecessary functionlity for the linux kernel in the resulting binary
image.