[configs] strip kernel exports
This should reduce the size of the kernel and the resulting firmware image. The
current configuration checks a build hash not just the kernel version so it's
nearly impossible to install other kernel modules than the modules that are
part of the release. This should not make the situation worse.

Change was proposed by Patrick Grimm.