Makefile: ensure quilt push -a can't fail when already on top of the patchset
`quilt push -a` fails if a patch is broken, but also fails when calling
it on a patch series which is already applied. For this reason we can
not ignore the return code of `quilt push -a`.
Calling `quilt pop -a` and ignore it's return code before calling
`quilt push -a` fix this problem.