change syntax of MAINTARGET-SUBTARGET naming
In commit "append subtarget name to all platforms" we switched to
a syntax of $TARGET/$SUBTARGET for all targets we support.

This changes the syntax between Target and Subtarget to "-". So
it's in the same way  as used in OpenWrt and LEDE and also avoids
doube "_", which is officially used by some subtargets.