assemble_firmware: workaround for failing imagebuilder when "BIN_DIR" is defined
It came up that the imagebuilder is failing for some Buffalo-routers when
BIN_DIR is supplied as make option. (see https://github.com/freifunk-berlin/firmware/pull/434)
This Hotfix changes the make-call to not use "BIN_DIR" and is doing the
file-moving separately.

When this is fixed upstream, this commit can simply be reverted to
have it nice looking again