Set "umask 022" for "git clone" and "feed update"
If the openwrt files are fetched using "git" on systems with "umask 002",
they get checked out with group-writable permission. This finally leads
to a "500 internal server error" when accessing luci pages, because of
wrong permissions of /usr/share/acl.d/luci-base.json.

See OpenWrt: "Builds with umask != 022 are known to be broken":
https://github.com/openwrt/luci/issues/1521#issuecomment-356649476

Bugreport and details, how these wrong permissions end up in the image:
https://github.com/freifunk-berlin/firmware/issues/431