Update to OpenVPN 2.4.5
Reenable compile directives for the LZO library for OpenVPN 2.4.5.  The core
problem with version 2.4.5 is that the comp.lzo option is not longer
depricated, but now totally removed.  As shown in issue 580, the only
working option found to date is to reenable building with LZO support
and to use the "compress lzo" option.