Makefile: add new target "initrd"
For several boards an inited-image is needed to do the initial bootup
of OpenWrt. When this image is booted the board can be finally flashed
to OpenWrt.
As such initrd-images will be only built by compiling OpenWrt we can't
use the imagebuilder. This target copies the initrd-filed into the root
of the resulting firmware-directory.

Also move the section to copy the mikrotik-inird into this section.