README: link to the wiki-page for enduser relevant informations (#655)
* README: link to the wiki-page for the a description of the image-types

As the Wiki is more up to date in this regard, just put a link in the Readme

* Readme: strip down to information relevant to developers

Rely on the firmware-pages in the Freifunk-Wiki for end-user relevant information (ReleaseNotes, Features, ...). So these information do not need to be maintained in two places. Also people interested in development will for sure know about setup and the recent features of the software and like to quickly read how to get involved.

* README: fix link to Setup-Tutorial