Makefile: update feeds after applying patches
When a package is broken as of a problem in its Makefile it will not
be installed by "scripts/feeds install". When we apply a patch to fix this
the corrected Makefile will not be checked again. As a result the package
will not be build, as it's still not installed.
This fixes the issue by calling "scripts/feeds" script again.