configs: build with MULTI_PROFILE & ALL_PROFILES
In issue #738 we found the problem, that several profiles can't
be build with the imagebuilder, as board-specific drivers will
be missing.
For the official supported boards we fixed this by enabling such
drivers manually in the targets config-file. But for users who
want to use the imagebuilder to generate an firmware-image for
an individual board this hardware-module will be missing. Doing
a full compiler-run was the only solution here.
When we set the CONFIG_ALL_KMODS, which build all kernel-modules,
this solved the problem along with having all kernel-modules.

The proper way to have all profile relevant modules build is
setting CONFIG_TARGET_MULTI_PROFILE and CONFIG_TARGET_ALL_PROFILES.
This has also the advantage, that we do not depend on ALL_KMODS.

Based on this config-change we can drop the explicit kernel-modules
for board-support we added in the target-configs.

Cherry-picked from: 04c3b4f96d7a0d1ceeeaa79f08c3b9e061bbe5b7