scripts: do not create gluon_base entry in feeds.conf
The gluon-script "feeds.sh" creates by default a "gluon_base" entry in the feeds.conf
of OpenWrt. This links to the gluon-base packages directory and causes an error message
during build, as this directory does not exist in our firmware.
Add a test for definiton of a "FOREIGN_BUILD" variable to the feeds.sh script, which
will no add the mentioned line to the feeds.conf, when defined.