rename packages folder to packagelists
When we started to use some code of the gluon framework a minor conflict of the usage
of the packages-folder came up.
We use it to store the package-lists for our images, Gluon uses this folder to store
the upstream repos. By this double-use there are always the repos inside the folder
and causing git to report them as untracked files. Adding the folder to .gitignore is
also no option, as we need it for our package-lists.
So just lets move our files and adjust .gitignore.