fix creating of VERSION.txt
Since we apply the patches on the external repos, instead of just patching
the upstream repo (e81f4109c) our VERSION.txt file shows this applied patch
as commit of the repo. But the intention of the file is to show the original
upstream commit we base on. So we need to use the "base" branch created by
the patch.sh script.
To ease maintenance of the function, we move it into a separate script, that
we call from the Makefile.