build: use separate Makefile for CI and drop IS_BUILDBOT
Add a separate Makefile that will be used for building in a CI-environment.
This new Makefile.autobuild adopts some simple make-targets extracted from
the OpenWrt-buildbot configuration, as our current "real" Makefile is not
atomic enough.
The new file will also directly deal with the special cases we control with
the "IS_BUILDBOT" and the "SET_BUILDBOT" envs / flags. So we can get rid of
them again and have two more easy to read Makefiles without additional
"#if" macros.

makefile.autobuild:
* reuse initialization from "real" Makefile
* common initialization becomes goal openwrt/.config
 * depend on configs/common.autobuild
 * add direct target "config"
* targets for individual CI-systems (https://github.com/freifunk-berlin/firmware/issues/804)
  * "buildbot" which will behave like the current buildbot setup calling
    the monolithic "Real" Makefile
  * "travis" has been used to for initial test on TravisCI, but this was
    a dead road
* the simple targets copied from OpenWrt are based on their 004f75d832c73b1e63f8d0ce5537d29e327fbc5b

Makefile, config.mk:
* remove "SET_BUILDBOT" and "IS_BUILDBOT" related stuff