GHA/pr-conflict: add workflow to automatically find conflicting PRs
Adds an GithubAction that checks if any open PR can not be merged on master
branch after a new commit was made to the master branch. If such conflict
will be detected, the affected PR will get the label "has conflicts". If
during a following run a conflicting PR becomes mergeable, the flag will be
removed again.
This action is based on https://github.com/marketplace/actions/auto-label-merge-conflicts