Misc fixes without functional change for enduser (#807)
* profiles: drop mt7620a eval-board from ramips-mt7620

This is another evaluation board that we create images for. It was added in
5f5122ca60408bbf70e5f461ffdddc6b3a6cc45c, when switching from "Default" board
list to explicit board definitions. during this change it seems that just all
boards have been added.
Our statistics don't list any running board, so just disable it.

* .gitignore: fix openwrt/ and tmp/

This folders are still listed as unversioned files, even they are not
relevant for us.

* profiles: drop Ubnt LS-SR71

This is an embedded board for a miniPCI socket, which was added when switching
to LEDE buildbasis (a05e97b6ac1333b0c710d6de65f762ee57642a07). As we focus on
regular off-the-shelf routers, we stop to build this image.
On the statistics there is also no indication, that this board is used anyway.

* README: remove explicit requirement for quilt

Since cfae54922054d98138165d99faf004566535035b the usage of quilt has been replaced
by 'git am'.