patches/openwrt: move 8MB Flash / 32MB RAM boards to ath79-tiny
Even these boards have enough flash to install the full list of packages, they are faced
frequent OoM-errors. So move them to the tiny target, to benefit from the optimized settings.
This change was discussed Upstream [1], but decided to be not the generic usecase that OpenWrt will follow.

The most relevant boards affected are the Ubiquiti XM plattform (Nanostation M), TL-WR1043N/ND v1 and TL-WR842N/ND v1/v2.

1 - http://lists.openwrt.org/pipermail/openwrt-devel/2020-September/031422.html