packagelists: include "luci-proto-ipv6" to allow editing wan6-interface
Without this package the "wan6" interface shows "unknown protocol" in LuCI, so
including it manually. OpenWrt uses the "luci" metapackage to get this included,
but we don't want the full LuCI.
Adding a dependency to luci-mod-network is not accepted upstream for policy
reasons (see https://github.com/openwrt/luci/pull/5198).